Move some resuable proofchecking functions
