+++ Carnap/Carnap.cabal
@@ -82,6 +82,7 @@ library
                      , Carnap.Languages.SetTheory.Logic.Carnap
                      , Carnap.Languages.ClassicalSequent.Syntax
                      , Carnap.Languages.ClassicalSequent.Parser
+                     , Carnap.Calculi.Util
                      , Carnap.Calculi.NaturalDeduction.Checker
                      , Carnap.Calculi.NaturalDeduction.Parser
                      , Carnap.Calculi.NaturalDeduction.Parser.Util
